Partial revert of some model orientation changes due to code issue.
Found Eagle-Loader Connectors on the back-side of the board weren't
orientated correctly. This lead to me committing some orientation
changes which broke correct display of the same components on the
*top* side of the board.

To get the benefit of this fix you need a corrected version of the
library (not yet committed).

Partial revert of 9bcf96f56ddb22e3da5861ff4540a5e6bb6936ab.